Added input validation to web api

+++ WarehouseApp/Services/WarehouseService.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using Microsoft.AspNetCore.Mvc.ModelBinding;
+using Microsoft.EntityFrameworkCore;
 using System.Text.RegularExpressions;
 using WarehouseWebApi.Mappers;
 using WarehouseWebApi.Repository;
@@ -17,14 +18,24 @@ namespace WarehouseWebApi.Services
             _repository = repository;
             _mapper = ServiceMapperConfig.InitializeAutomapper();
         }
-        
+
         internal async Task AddPallet(Models.View.Pallet pallet, CancellationToken cancellationToken)
         {
             var palletBusiness = _mapper.Map<Models.Business.Pallet>(pallet);
 
             BuidDataIntegrity(palletBusiness.Boxes, null, palletBusiness);
 
-            await _repository.AddPallet(palletBusiness, cancellationToken);
+            try
+            {
+                await _repository.AddPallet(palletBusiness, cancellationToken);
+            }
+            catch (DbUpdateException exception)
+            {
+                if (exception?.InnerException is Npgsql.PostgresException innerException && innerException.SqlState == "23505")
+                {
+                    throw new InvalidDataException("Operation failed due to existing in warehouse barcode.");
+                }
+            }
         }
 
         internal async Task<Models.View.Pallet[]> GetPallets(CancellationToken token)
@@ -47,6 +58,13 @@ namespace WarehouseWebApi.Services
 
         internal async Task RemoveBoxes(string[] barcodes, CancellationToken token)
         {
+            var availableBarcodes = await _repository.GetBoxesBarcodes(token);
+
+            if (availableBarcodes.Intersect(barcodes).Count() != barcodes.Length)
+            {
+                throw new InvalidDataException("Operation denied due to unexisting in warehouse barcode.");
+            }
+
             await _repository.RemoveBoxes(barcodes, token);
         }
 
@@ -63,12 +81,42 @@ namespace WarehouseWebApi.Services
                     if (!m.Success)
                     {
                         modelState.AddModelError("barcode", $"'{barcode}' is not alpha numeric");
-                        return true;
                     }
                 }
             }
 
-            return false;
+            return modelState.ErrorCount > 0;
+        }
+
+        internal static bool IsInputInvalid(Models.View.Pallet pallet, ModelStateDictionary modelState)
+        {
+            var barcodes = CollectBarcodes(pallet.Boxes).ToArray();
+
+            if ( barcodes.Distinct().Count() < barcodes.Length)
+            {
+                modelState.AddModelError($"pallet: {pallet.Barcode}", $"contains duplicate barcodes.");
+            }
+
+            HasInvalidBarcodes(barcodes, modelState);
+
+            return modelState.ErrorCount > 0;
+        }
+
+        private static IEnumerable<string> CollectBarcodes(IEnumerable<Models.View.Box> boxes)
+        {
+            IEnumerable<string> barcodes = Enumerable.Empty<string>();
+
+            boxes ??= Enumerable.Empty<Models.View.Box>();
+
+            foreach (var box in boxes)
+            {
+                barcodes = barcodes.Append(box.Barcode);
+
+                var subCodes = CollectBarcodes(box.Boxes);
+
+                barcodes = barcodes.Concat(subCodes);
+            }
+            return barcodes;
         }
     }
 }
